Move user preferences from shared to client
As part of the move toward a stateless server session, the
IPreferences, IUserPreferencesService, IUserPreferencesStorageService,
IPreferenceChangeListener and their corresponding implementations
have been moved from shared to client

423389
