+++ org.eclipse.scout.rt.client/src/main/java/org/eclipse/scout/rt/client/services/common/prefs/FileSystemUserPreferencesStorageService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2023 BSI Business Systems Integration AG
+ * Copyright (c) 2010, 2025 BSI Business Systems Integration AG
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -32,8 +32,6 @@ import org.eclipse.scout.rt.platform.BEANS;
 import org.eclipse.scout.rt.platform.config.ConfigUtility;
 import org.eclipse.scout.rt.platform.config.IConfigProperty;
 import org.eclipse.scout.rt.platform.exception.ProcessingException;
-import org.eclipse.scout.rt.shared.services.common.prefs.AbstractUserPreferencesStorageService;
-import org.eclipse.scout.rt.shared.services.common.prefs.IPreferences;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
